Refactor BatchDelete into an outer class
